# spring-test/src/main/resources/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sample API with Different Responses",
    "version": "v1"
  },
  "paths": {
    "/users": {
      "get": {
        "summary": "Get all users",
        "responses": {
          "200": {
            "description": "A list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "Get a product by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID of the product to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
    "/orders": {
      "post": {
        "summary": "Create a new order",
        "requestBody": {
          "description": "Order to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "required": ["id", "name", "email"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "Product": {
        "type": "object",
        "required": ["id", "name", "price"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "OrderInput": {
        "type": "object",
        "required": ["user_id", "product_id", "quantity"],
        "properties": {
          "user_id": {
            "type": "integer"
          },
          "product_id": {
            "type": "integer"
          },
          "quantity": {
            "type": "integer"
          }
        }
      },
      "Order": {
        "type": "object",
        "required": ["id", "user_id", "product_id", "quantity", "total_amount"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "product_id": {
            "type": "integer"
          },
          "quantity": {
            "type": "integer"
          },
          "total_amount": {
            "type": "number",
            "format": "float"
          }
        }
      }
    }
  }
}